adding preliminary vagrant implementation
These changes provide a vagrant box that will have postgresql installed
and running. There's still bootstrapping code that needs to be added but
this implementation does work with a properly configured laptop
(bootstrap code coming in a bit). I'm checking in the thirdparty
cookbooks for now, and will added Librarian in the future manage them
that way.