adding vagrant bootstrapper
This script assumes system ruby usage. Will need to test it with rvm and
figure out how to handle that scenario. Attempting to do as much ruby
stuff in the VM as possible.

This script installs chef in system ruby and uses it to bootstrap your
host operating system (current only works with OSX). It will download
and install the VirtualBox.dmg and Vagrant.dmg

Potential issues are with the fact that the dmg installer doesn't care
if the versions of these tools are old, so it happily continues going
when it should be a little more opinionated on what versions are on your
machine to ensure conformity.