Regarding the FB meta tag og:image on the petition page...
For some (unknown to me) reason, VK was always using the first
default facebook share image if the visitor was seeing a petition
for the first time or had not signed it. In this case it would not
use the petition image. Now it will always use a petition image
if one is available, else it will choose a default.